services/order/requirements/base.txt
amqp==2.5.2
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.26
django-filter==1.1.0
django-redis==4.11.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
idna==2.8
importlib-metadata==1.3.0
itypes==1.1.0
Jinja2==2.10.3
kombu==4.6.7
MarkupSafe==1.1.1
more-itertools==8.0.2
openapi-codec==1.3.2
python-dateutil==2.8.1
pytz==2019.3
redis==3.3.11
requests==2.22.0
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.25.7
vine==1.3.0
zipp==0.6.0
psycopg2==2.8.4
djangorestframework-csv==2.1.0
unicodecsv==0.14.1
